Refactor merchant API
