+++ example/package.json
@@ -14,6 +14,7 @@
     "@react-navigation/native-stack": "^6.1.0",
     "@types/react": "^17.0.24",
     "@types/react-native": "^0.65.2",
+    "axios": "^0.26.1",
     "react": "16.13.1",
     "react-native": "0.63.4",
     "react-native-safe-area-context": "^3.3.2",
